% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cval.R
\name{as.data.frame.clubprorand}
\alias{as.data.frame.clubprorand}
\title{Convert the output of pcc_replicates() to a data.frame.}
\usage{
\method{as.data.frame}{clubprorand}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{an object of class "clubprorand"}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{...}{ignored}
}
\description{
Convert the output of pcc_replicates() to a data.frame.
}
\details{
This function is useful to format pcc replicates data for plotting.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
z <- pcc_replicates(mod)
as.data.frame(z)
}
