% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caffeine}
\alias{caffeine}
\title{Caffeine data}
\format{
A data frame with 30 rows and 2 columns:
\describe{
 \item{dose}{dose of caffeine in mg}
 \item{rate}{finger taps per minute}
}
}
\source{
Hand D.J., Daly F., Lunn A.D., McConway K.J., Ostrowski E. (1994) A Handbook of Small Data Sets. London: Chapman & Hall. Data set 50.
}
\usage{
caffeine
}
\description{
Effect of three different doses of caffeine on finger tapping rate.
}
\keyword{datasets}
