% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify observations}
\usage{
classify(y, x)
}
\arguments{
\item{y, }{a numeric vector of observations.}

\item{x, }{a factor vector.}
}
\value{
an object of class "clubprofit" is a list containing the folllowing
components:
\describe{
  \item{prediction}{a character vector of predicted classifications.}
  \item{accuracy}{a character vector indicating whether each classification
  is "correct", "incorrect", or "ambiguous".}
  \item{pcc}{the percentage of correct classifications.}
  }
}
\description{
\code{classify()} is used to classify obervations using binary procrustes
rotation.
}
