% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare models.}
\usage{
compare(m1, m2)
}
\arguments{
\item{m1}{an object of class "clubprofit" produced by \code{club()}}

\item{m2}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
an object of type "clubprocomparison"
}
\description{
Compare models.
}
\details{
Compare the PCC of two clubprofit models and compute the chance-value of the
difference.
}
\examples{
m1 <- club(width ~ location, jellyfish)
m2 <- club(length ~ location, jellyfish)
compare(m1, m2)
}
