% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R
\name{csi}
\alias{csi}
\title{Classification strength indices.}
\usage{
csi(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric vector.
}
\description{
Classification strength indices.
}
\details{
Returns a vector containing the classification strength
index for each observation.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
csi(mod)
}
