% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cval}
\alias{cval}
\title{Chance value.}
\usage{
cval(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric value.
}
\description{
Chance value.
}
\details{
Compute the chance that randomly reordered data results in a
percentage of correctly classified observations at least as high as
the observed data.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- club(a, b)
cval(mod)
}
