% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{individual_results}
\alias{individual_results}
\title{Individual level classification results.}
\usage{
individual_results(m, digits)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{classify()}}

\item{digits}{an integer}
}
\value{
a data.frame containing a columns of predictions and prediction
accuracy
}
\description{
Individual level classification results.
}
\details{
Returns a data.frame containing predicted classifications and classification
accuracy for each individual observation.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- club(a, b)
individual_results(mod)
individual_results(mod, digits = 3)
}
