% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jellyfish}
\alias{jellyfish}
\title{Jellyfish dimension data}
\format{
## `jellyfish`
A data frame with 46 rows and 3 columns:
\describe{
 \item{location}{location where jellyfish was caught}
 \item{width}{jellyfish width in mm}
 \item{length}{jellyfish length in mm}
}
}
\source{
Hand D.J., Daly F., Lunn A.D., McConway K.J., Ostrowski E. (1994) A Handbook of Small Data Sets. London: Chapman & Hall. Data set 225.
}
\usage{
jellyfish
}
\description{
Sizes of jellyfish from two locations in the Hawkesbury River, New South Wales, Australia.
}
\keyword{datasets}
