% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{n_ambiguous}
\alias{n_ambiguous}
\title{Number of ambiguous classifications.}
\usage{
n_ambiguous(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
an integer.
}
\description{
Number of ambiguous classifications.
}
\details{
Returns the number of observations which were
classified ambiguously by the model.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
n_ambiguous(mod)
}
