% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{n_correct}
\alias{n_correct}
\title{Number of correct classifications.}
\usage{
n_correct(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
an integer.
}
\description{
Number of correct classifications.
}
\details{
Returns the number of observations which were
classified correctly by the model.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
n_correct(mod)
}
