% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pcc}
\alias{pcc}
\title{Percentage of correct classifications.}
\usage{
pcc(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric value.
}
\description{
Percentage of correct classifications.
}
\details{
Returns the percentage of correctly classified observations.
}
\examples{
df <- data.frame(a = sample(1:5, 20, replace = TRUE),
                 b = rep(c("group1", "group2"), each = 10))
df$b <- factor(df$b)
mod <- club(a ~ b, df)
pcc(mod)
}
