% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pcc_replicates}
\alias{pcc_replicates}
\title{PCC replicates.}
\usage{
pcc_replicates(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
an object of class clubprorand.
}
\description{
PCC replicates.
}
\details{
Returns an object containing a vector of PCC replicates used to
calculate the chance-value.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- club(a, b)
pcc_replicates(mod)
}
