% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{plot.clubproaccuracy}
\alias{plot.clubproaccuracy}
\title{Plot accuracy.}
\usage{
\method{plot}{clubproaccuracy}(x, ...)
}
\arguments{
\item{x}{an object of class "clubproaccuracy"}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Plot accuracy.
}
\details{
Produces a mosaic plot of predictio naccuracy by category
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
z <- accuracy(mod)
plot(z)
}
