% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{plot.clubprocomparison}
\alias{plot.clubprocomparison}
\title{Plot model comparison.}
\usage{
\method{plot}{clubprocomparison}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprocomparison".}

\item{...}{ignored}
}
\value{
no return value, called for side effects only.
}
\description{
Plot model comparison.
}
\details{
Plot a distribution of PCCs computed from randomly reordered data
used to calculate the chance-value for a model comparison.
}
\examples{
m1 <- club(width ~ location, jellyfish)
m2 <- club(length ~ location, jellyfish)
z <- compare(m1, m2)
plot(z)
}
