% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csi.R
\name{plot.clubprocsi}
\alias{plot.clubprocsi}
\title{Plot classification strength indices.}
\usage{
\method{plot}{clubprocsi}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprocsi"}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Plot classification strength indices.
}
\details{
Produces dotplot showing classification strength for each individual.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
z <- csi(mod)
plot(z)
}
