% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.clubprofit}
\alias{plot.clubprofit}
\title{Plot classification results.}
\usage{
\method{plot}{clubprofit}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprofit" produced by \code{classify()}}

\item{...}{ignored}
}
\value{
an object of class "ggplot"
}
\description{
Plot classification results.
}
\details{
Produces a bar plot of counts of individuals against observed values within
each possible classification grouping. Bars fill colours indicate whether
individuals were classified correctly, incorrectly or ambiguously.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- classify(a, b)
plot(mod)
}
