% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.clubpropredictions}
\alias{plot.clubpropredictions}
\title{Plot predictions.}
\usage{
\method{plot}{clubpropredictions}(x, ...)
}
\arguments{
\item{x}{an object of class "clubpropredictions"}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Plot predictions.
}
\details{
Produces a mosaic plot of observed versus predicted categories
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
z <- predict(mod)
plot(z)
}
