% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{plot.clubprothreshold}
\alias{plot.clubprothreshold}
\title{Plot PCC as a function of binary category boundary location.}
\usage{
\method{plot}{clubprothreshold}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprothreshold"}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Plot PCC as a function of binary category boundary location.
}
\details{
Produces an xyplot showing the PCC returned for each possible category boundary.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
z <- threshold(mod)
plot(z)
}
