% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_classification}
\alias{plot_classification}
\alias{plot_classification.default}
\alias{plot_classification.clubprofit}
\title{Plot classification accuracy.}
\usage{
plot_classification(m, ...)

\method{plot_classification}{default}(m, ...)

\method{plot_classification}{clubprofit}(m, abline_offset = -0.1, point_offset = 0.4, ...)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{classify()}}

\item{...}{ignored}

\item{abline_offset}{a number indicating vertical adjustment of horizontal ablines}

\item{point_offset}{a number indicating amount of vertical separation between points}
}
\value{
called for side-effects only
}
\description{
Plot classification accuracy.
}
\details{
Produces a strip plot showing counts of individuals against observed values within
each target grouping. Point fill colours indicate whether
individuals were classified correctly, incorrectly or ambiguously.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- classify(a, b)
plot_classification(mod)
}
