% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_prediction}
\alias{plot_prediction}
\alias{plot_prediction.default}
\title{Plot predicted classification results.}
\usage{
plot_prediction(m, ...)

plot_prediction.default(m, ...)

plot_prediction(m, ...)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{classify()}}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Plot predicted classification results.
}
\details{
Produces a strip plot showing counts of individuals for each observed value by
predicted classification grouping.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- classify(a, b)
plot_prediction(mod)
}
