% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{print.clubproaccuracy}
\alias{print.clubproaccuracy}
\title{Print prediction accuracy.}
\usage{
\method{print}{clubproaccuracy}(x, ...)
}
\arguments{
\item{x}{an object of class "clubproaccuracy"}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Print prediction accuracy.
}
\details{
Print a table of prediction accuracy by category
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
z <- accuracy(mod)
print(z)
}
