% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{print.clubprofit}
\alias{print.clubprofit}
\title{Print the model call.}
\usage{
\method{print}{clubprofit}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprofit"}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Print the model call.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
print(mod)
}
