% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{print.clubpropredictions}
\alias{print.clubpropredictions}
\title{Print predictions.}
\usage{
\method{print}{clubpropredictions}(x, ...)
}
\arguments{
\item{x}{an object of class "clubpropredictions"}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Print predictions.
}
\details{
Print a table of observed versus predicted categories
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
z <- predict(mod)
print(z)
}
