% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{summary.clubprocomparison}
\alias{summary.clubprocomparison}
\title{Generate a summary of a comparison of clubprofit models.}
\usage{
\method{summary}{clubprocomparison}(object, ...)
}
\arguments{
\item{object}{an object of class "clubprocomparison".}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Generate a summary of a comparison of clubprofit models.
}
\examples{
m1 <- club(width ~ location, jellyfish)
m2 <- club(length ~ location, jellyfish)
z <- compare(m1, m2)
summary(z)
}
