% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{summary.clubprofit}
\alias{summary.clubprofit}
\title{Print a summary of results from a fitted clubpro model.}
\usage{
\method{summary}{clubprofit}(object, ..., digits = 2L)
}
\arguments{
\item{object}{an object of class "clubprofit".}

\item{...}{ignored}

\item{digits}{an integer used for rounding numeric values in the output.}
}
\value{
No return value, called for side effects.
}
\description{
Print a summary of results from a fitted clubpro model.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- classify(a, b)
summary(mod)
summary(mod, digits = 3)
}
