% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Classification strength indices.}
\usage{
threshold(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
an object of class clubprothreshold
}
\description{
Classification strength indices.
}
\details{
Returns a vector containing the classification strength
index for each observation.
}
\examples{
mod <- club(width ~ location, data = jellyfish)
threshold(mod)
}
