% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{accuracy}
\alias{accuracy}
\title{Classification accuracy for each observation.}
\usage{
accuracy(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a table
}
\description{
Classification accuracy for each observation.
}
\details{
Returns a character vector containing a string corresponding to each
observation indicating whether classification of that observation was
"correct", "incorrect", or "ambigous".
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
accuracy(mod)
}
