% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classify observations}
\usage{
classify(y, x, imprecision = 0, nreps = 1000L, normalise_cols = TRUE)
}
\arguments{
\item{y}{a factor vector of observations.}

\item{x}{a factor vector of target groups.}

\item{imprecision}{a number indicting the margin of imprecision allowed in classification.}

\item{nreps}{the number of replicates to use in the randomisation test.}

\item{normalise_cols}{a boolean indicating whether to normalise matrix columns.}
}
\value{
an object of class "clubprofit" is a list containing the folllowing
components:
\describe{
  \item{prediction}{a character vector of predicted classifications.}
  \item{accuracy}{a character vector indicating whether each classification
  is "correct", "incorrect", or "ambiguous".}
  \item{pcc}{the percentage of correct classifications.}
  \item{cval}{the chance of randomly reordered data producing a PCC >= the
  observed PCC.}
  \item{pcc_replicates}{a vector of PCCs generated from randomly reordered
  data used to calculate \code{cval}.}
  \item{call}{the matched call.}
  }
}
\description{
\code{classify()} is used to classify obervations using binary procrustes
rotation.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- classify(a, b)
mod <- classify(a, b, nreps = 200L)
}
