% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{csi}
\alias{csi}
\title{Classification strength indices.}
\usage{
csi(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric vector.
}
\description{
Classification strength indices.
}
\details{
Returns a vector containing the classification strength
index for each observation.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- club(a, b)
csi(mod)
}
