% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cval.R
\name{cval}
\alias{cval}
\title{Chance value.}
\usage{
cval(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric value.
}
\description{
Chance value.
}
\details{
Compute the chance that randomly reordered data results in a
percentage of correctly classified observations at least as high as
the observed data.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
cval(mod)
}
