% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{median_csi}
\alias{median_csi}
\title{Median classification strength index.}
\usage{
median_csi(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric vector.
}
\description{
Median classification strength index.
}
\details{
Returns the median classification strength index.
}
\examples{
df <- data.frame(a = sample(1:5, 20, replace = TRUE),
                 b = rep(c("group1", "group2"), each = 10))
df$b <- factor(df$b)
mod <- club(a ~ b, df)
median_csi(mod)
}
