% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{median_csi}
\alias{median_csi}
\title{Median classification strength index.}
\usage{
median_csi(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric vector.
}
\description{
Median classification strength index.
}
\details{
Returns the median classification strength index.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
median_csi(mod)
}
