% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{n_correct}
\alias{n_correct}
\title{Number of correct classifications.}
\usage{
n_correct(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
an integer.
}
\description{
Number of correct classifications.
}
\details{
Returns the number of observations which were
classified correctly by the model.
}
\examples{
df <- data.frame(a = sample(1:5, 20, replace = TRUE),
                 b = rep(c("group1", "group2"), each = 10))
df$b <- factor(df$b)
mod <- club(a ~ b, df)
n_correct(mod)
}
