% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pcc}
\alias{pcc}
\title{Percentage of correct classifications.}
\usage{
pcc(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{club()}}
}
\value{
a numeric value.
}
\description{
Percentage of correct classifications.
}
\details{
Returns the percentage of correctly classified observations.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
pcc(mod)
}
