% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.clubprofit}
\alias{plot.clubprofit}
\title{Plot classification accuracy.}
\usage{
\method{plot}{clubprofit}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprofit" produced by \code{club()}}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Plot classification accuracy.
}
\details{
Produces bar plot showing counts of individuals against observed values within
each target grouping. Fill colours indicate whether each individual was
classified correctly, incorrectly or ambiguously.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- club(a, b)
plot(mod)
}
