% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.clubprofit}
\alias{plot.clubprofit}
\title{Plots individual-level classification results.}
\usage{
\method{plot}{clubprofit}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprofit" produced by \code{classify()}}

\item{...}{ignored}
}
\value{
an object of class "ggplot"
}
\description{
Plots individual-level classification results.
}
