% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot.clubprorand}
\alias{plot.clubprorand}
\title{Plot PCC replicates.}
\usage{
\method{plot}{clubprorand}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprofit" produced by \code{club()}}

\item{...}{ignored}
}
\value{
no return value, called for side effects only.
}
\description{
Plot PCC replicates.
}
\details{
Plot the distribution of PCCs computed from randomly reordered data
used to calculate the chance-value.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
plot(pcc_replicates(mod))
}
