% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_classification}
\alias{plot_classification}
\alias{plot_classification.default}
\alias{plot_classification.clubprofit}
\title{Plot classification accuracy.}
\usage{
plot_classification(x, ...)

\method{plot_classification}{default}(x, ...)

\method{plot_classification}{clubprofit}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprofit" produced by \code{classify()}}

\item{...}{ignored}
}
\value{
called for side-effects only
}
\description{
Plot classification accuracy.
}
\details{
Produces a strip plot showing counts of individuals against observed values within
each possible classification grouping. Bars fill colours indicate whether
individuals were classified correctly, incorrectly or ambiguously.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- classify(a, b)
plot_classification(mod)
}
