% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{plot_cvalue}
\alias{plot_cvalue}
\alias{plot_cvalue.default}
\alias{plot_cvalue.clubprofit}
\title{Plot observed and random order generated PCCs.}
\usage{
plot_cvalue(m)

\method{plot_cvalue}{default}(m)

\method{plot_cvalue}{clubprofit}(m)
}
\arguments{
\item{m}{an object of class "clubprofit" produced by \code{classify()}.}
}
\value{
an object of class "ggplot".
}
\description{
Produces a plot which shows the proportion of random reorderings of the
observed data which produce PCCs at least as high as the observed PCC.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- classify(a, b)
plot_cvalue(mod)
}
