% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.clubprofit}
\alias{predict.clubprofit}
\title{Predicted category for each observation.}
\usage{
\method{predict}{clubprofit}(object, ...)
}
\arguments{
\item{object}{an object of class "clubprofit" produced by \code{club()}}

\item{...}{ignored}
}
\value{
a table
}
\description{
Predicted category for each observation.
}
\details{
Returns a character vector containing the name of the predicted category
for each observed value.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
predict(mod)
}
