% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{predict.clubprofit}
\alias{predict.clubprofit}
\title{Predicted category for each observation.}
\usage{
\method{predict}{clubprofit}(object, ...)
}
\arguments{
\item{object}{an object of class "clubprofit" produced by \code{club()}}

\item{...}{ignored}
}
\value{
a character vector.
}
\description{
Predicted category for each observation.
}
\details{
Returns a character vector containing the name of the predicted category
for each observed value.
}
\examples{
df <- data.frame(a = sample(1:5, 20, replace = TRUE),
                 b = rep(c("group1", "group2"), each = 10))
df$b <- factor(df$b)
mod <- club(a ~ b, df)
predict(mod)
}
