% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{print.clubprofit}
\alias{print.clubprofit}
\title{Print the model call.}
\usage{
\method{print}{clubprofit}(x, ...)
}
\arguments{
\item{x}{an object of class "clubprofit"}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Print the model call.
}
\examples{
a <- sample(1:5, 20, replace = TRUE)
b <- rep(c("group1", "group2"), each = 10)
b <- factor(b)
mod <- club(a, b)
print(mod)
}
