% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.clubprofit}
\alias{summary.clubprofit}
\title{Generate a summary of results from a fitted clubpro model.}
\usage{
\method{summary}{clubprofit}(object, ...)
}
\arguments{
\item{object}{an object of class "clubprofit".}

\item{...}{ignored}
}
\value{
No return value, called for side effects.
}
\description{
Generate a summary of results from a fitted clubpro model.
}
\examples{
mod <- club(rate ~ dose, data = caffeine)
summary(mod)
}
